Cleaning up stuff and finalizing design/architecture
Also updated README.md to reflect how to build/run/configure the app.
It now has a shadowed jar version!  wohoo!
Realized I missed the part of the reqs where clickTimer comes in as a property, so moved that from yml file to system property.
I am in fact committing a password, although its solely for a local/dummy database. Please change this if you actually use it.
For saving the Auctions, I needed to manually bind to a hibernate session, so we have a bit of a funky way to that. In real world should be done via a better way
Decided to use the concept of an empty bid that has a provider id of -1.  This will be an issue if the database it's stood up against is missing this. However this means every Auction has at least one bid, which is very convenient.
Renamed AdResource to AuctionResource as that's what it really is. Allows me to treat Auctions just like Users, Providers, and AdSizes (the other primary entities)
Added a basic unit test around auctioning.  It's difficult to test performance with so much mocking and introspection going on, but an Integrationn test in a fleshed out CI environment would help with that. The other functionss are pretty basic, and writing unit tests for passthrough/getter/setter methods seem silly. May add some for registerClick later.